backend/pyproject.toml
[project]
name = "dkmes-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.123.5,<0.124.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "google-cloud-aiplatform (>=1.129.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "falkordb (>=1.2.0,<2.0.0)",
    "chromadb (>=1.3.5,<2.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pypdf (>=6.4.0,<7.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
